third_party/WebKit/LayoutTests/http/tests/media/media-source/mediasource-sourcebuffer-mode-expected.txt

CONSOLE WARNING: The provided value 'invalidmode' is not a valid enum value.
CONSOLE WARNING: The provided value 'null' is not a valid enum value.
CONSOLE WARNING: The provided value '' is not a valid enum value.
CONSOLE WARNING: The provided value 'Segments' is not a valid enum value.
 
PASS Test initial value of SourceBuffer.mode is "segments" 
PASS Test setting SourceBuffer.mode 
PASS Test setting a removed SourceBuffer's mode 
PASS Test setting SourceBuffer.mode while still updating 
PASS Test setting SourceBuffer.mode triggers parent MediaSource 'ended' to 'open' transition. 
PASS Test setting SourceBuffer.mode and SourceBuffer.timestampOffset while parsing media segment. 
